Added feature "setExpectedContentType" to CCHttpRequest
Added the ability to specify an expected content type string. When the CCHttpRequest is successful, a check of the expectedContentType is made against the incoming response. If the the response's content-type contains the the specified content-type string, the request is successful and the delegate is notified. If the content-types differ, the request is considered unsuccessful and the delegate is notified of the failure with an explanation of the difference in the error message.

If no expected content type is specified, the request will behave as before.